refactor(cli): rename internal/grayc to internal/driver to avoid naming collision
The Go wrapper package shared its name with grayc/, the C compiler source
directory. Renamed to "driver" which accurately describes its role as the
compiler driver layer. Also added *.stub to gitignore.
