+++ .gitignore
@@ -93,9 +93,10 @@ context/
 # Embedded compiler runtime — generated by `make build`.
 # The files exist so //go:embed resolves; their binary content is
 # never committed. Fresh clones get empty stubs from `make stubs`.
-/internal/grayc/runtime/grayc
-/internal/grayc/runtime/libgrayrt.a
-/internal/grayc/runtime/src/
+/internal/driver/runtime/grayc
+/internal/driver/runtime/libgrayrt.a
+/internal/driver/runtime/src/
+*.stub
 *.md
 .claude/
