+++ internal/driver/grayc.go
@@ -5,7 +5,7 @@
 // Copyright (c) 2025-Present Marshall A Burns
 // Licensed under the MIT License. See LICENSE for details.
 
-package grayc
+package driver
 
 import (
 	"fmt"
